+++ packages/terra-core-docs/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+* Changed
+  * Fixed Focus Issue & Expand/collapse state for `terra-toggle` examples.
+* Added 
+  * Added Label Example for `terra-toggle` examples.
+
 ## 1.18.0 - (January 9, 2023)
 
 * Changed